requirements.txt
chainlit>=1.0.0
llama-cpp-python>=0.3.2
chromadb>=0.4.22
sentence-transformers>=2.2.2
huggingface-hub>=0.20.0
pymupdf>=1.23.0
numpy>=1.24.0
torch>=2.0.0
Pillow>=10.0.0
tqdm>=4.66.0

